Clean up and simplify the Portfolio splat path's `beforeLoad` with a valid path helper
- Instead of directly calling `find` on the list of expected valid Portfolio path IDs
for path comparison in the `beforeLoad` func, it's cleaner and easier to test a
separate path validator helper func. In the long run, this makes it easier to
read and easier to add additional redirect checks later if needed.
- Destructuring the `beforeLoad` options prop is also helpful to grab only what's
needed for comparison sake, i.e. the `params.postId` string, though more can be
grabbed later if needs change.